- fixed compilation warning reported by MSVC
src\r_videoscale.cpp(105): warning C4244: 'initializing': conversion from 'float' to 'int', possible loss of data

# Conflicts:
#	src/r_videoscale.cpp

+++ src/r_videoscale.cpp
@@ -188,7 +188,7 @@ bool ViewportIsScaled43()
 
 void R_ShowCurrentScaling()
 {
-	int x1 = screen->GetWidth(), y1 = screen->GetHeight(), x2 = x1 * vid_scalefactor, y2 = y1 * vid_scalefactor;
+	int x1 = screen->GetWidth(), y1 = screen->GetHeight(), x2 = int(x1 * vid_scalefactor), y2 = int(y1 * vid_scalefactor);
 	Printf("Current Scale: %f\n", (float)(vid_scalefactor));
 	Printf("Real resolution: %i x %i\nEmulated resolution: %i x %i\n", x1, y1, x2, y2);
 }
